Shorten file names
