+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/direct_query_data_source_configuration/amazon_s3_datasource/configure_amazon_s3_data_source.tsx
