+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/direct_query_data_source_configuration/prometheus_data_source/review_prometheus_data_source.test.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 import { mount } from 'enzyme';
 import { MemoryRouter } from 'react-router-dom';
 import { act } from 'react-dom/test-utils';
-import { ReviewPrometheusDatasource } from './direct_query_review_prometheus_data_source';
+import { ReviewPrometheusDatasource } from './review_prometheus_data_source';
 import { AuthMethod } from '../../../constants';
 import { Role } from '../../../../types';
