+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/direct_query_data_source_configuration/prometheus_data_source/review_prometheus_data_source.tsx
